service/requirements.txt
Django==2.0.3
pymysql==0.8.0
djangorestframework==3.7.7
requests
redis-cache==0.1.3
django-redis-cache==1.7.1
hiredis==0.2.0
django-extensions==2.0.6